Use X-CSRF-Token header name more consistently, a few places still used moquiSessionToken for the header name so adding both for now (may deprecate in the future); add UserFacade.getLoginKey(float expireHours) method and implementation

+++ framework/src/main/groovy/org/moqui/impl/context/ExecutionContextFactoryImpl.groovy
@@ -944,9 +944,9 @@ class ExecutionContextFactoryImpl implements ExecutionContextFactory {
         MNode loginKeyNode = confXmlRoot.first("user-facade").first("login-key")
         return loginKeyNode.attribute("encrypt-hash-type") ?: "SHA-256"
     }
-    int getLoginKeyExpireHours() {
+    float getLoginKeyExpireHours() {
         MNode loginKeyNode = confXmlRoot.first("user-facade").first("login-key")
-        return (loginKeyNode.attribute("expire-hours") ?: "144") as int
+        return (loginKeyNode.attribute("expire-hours") ?: "144") as float
     }
 
     // ====================================================
